Move test and fix int types
