+++ src/spu/src/smart_stream/file_batch.rs
@@ -44,7 +44,7 @@ impl FileBatchIterator {
         Self {
             fd: slice.as_raw_fd(),
             offset,
-            end: offset + slice.len(),
+            end: offset + slice.len() as i64,
         }
     }
 }
@@ -121,5 +121,33 @@ impl Iterator for FileBatchIterator {
         debug!(file_offset = self.offset, "fbatch end");
 
         Some(Ok(FileBatch { batch, records }))
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use std::{fs::File, io::Write};
+    use std::env::temp_dir;
+    use std::os::unix::io::AsRawFd;
+
+    use fluvio_storage::config::DEFAULT_MAX_BATCH_SIZE;
+
+    use super::*;
+
+    #[test]
+    fn test_file() {
+        let path = temp_dir().join("pread.txt");
+        let mut file = File::create(&path).expect("create");
+        file.write_all(b"Hello, world!").expect("write");
+        file.sync_all().expect("flush");
+        drop(file);
+
+        let read_only = File::open(path).expect("open");
+        let fd = read_only.as_raw_fd();
+        let mut buf = vec![0; DEFAULT_MAX_BATCH_SIZE as usize];
+        // let mut buf = BytesMut::with_capacity(64);
+        let bytes_read = pread(fd, &mut buf, 1).expect("");
+        println!("bytes read: {}", bytes_read);
+        assert!(bytes_read > 2);
     }
 }
